added detectortable to reco

+++ map/DetectorTable.h
@@ -15,8 +15,8 @@ public:
 	void buildTable();
 	int getConnector(int channel){return mapConnector[channel];}
 	char getAxis(int channel){return mapAxis[channel];};
-	double getPitch(int channel){return mapPitch[channel];};
-	double getInter(int channel, int channelPerp = -1);   //{return mapInter[channel];};
+	float getPitch(int channel){return mapPitch[channel];};
+	float getInter(int channel, int channelPerp = -1);   //{return mapInter[channel];};
 	int getStripNb(int channel){return mapStripNb[channel];};
 	std::vector<int> getNeighbours(int channel){return mapNgh[channel];};
 
@@ -26,8 +26,8 @@ private:
 
 	std::map<int, int> mapConnector;
 	std::map<int, char> mapAxis;
-	std::map<int, double> mapPitch;
-	std::map<int, std::vector<double>> mapInter;
+	std::map<int, float> mapPitch;
+	std::map<int, std::vector<float>> mapInter;
 	std::map<int, int> mapStripNb;
 	std::map<int, std::vector<int>> mapNgh;
 };
